+++ infrastructure/foundation/hostedcluster/terraform/modules/iam/main.tf
@@ -1,22 +1,25 @@
 data "aws_caller_identity" "current" {}
 
-# Fetch S3 endpoint SSL certificate details for provider thumbprint
-data "tls_certificate" "oidc" {
-  url = "https://${var.oidc_bucket_domain}/${var.infra_id}"
+locals {
+  # OIDC issuer URL for THIS cluster uses the SHARED bucket, per-cluster sub-path
+  oidc_issuer_url                  = "https://${var.shared_oidc_bucket}.s3.${var.aws_region}.amazonaws.com/${var.infra_id}"
+  oidc_provider_url_without_schema = replace("https://${var.shared_oidc_bucket}.s3.${var.aws_region}.amazonaws.com/${var.infra_id}", "https://", "")
 }
 
-# IAM OIDC Identity Provider
+# Fetch S3 endpoint SSL certificate details for the OIDC provider thumbprint.
+# The TLS cert is for the regional S3 endpoint (e.g. s3.ap-southeast-1.amazonaws.com).
+data "tls_certificate" "oidc" {
+  url = "https://s3.${var.aws_region}.amazonaws.com"
+}
+
+# IAM OIDC Identity Provider for THIS cluster
 resource "aws_iam_openid_connect_provider" "oidc" {
-  url             = "https://${var.oidc_bucket_domain}/${var.infra_id}"
+  url             = "https://${var.shared_oidc_bucket}.s3.${var.aws_region}.amazonaws.com/${var.infra_id}"
   client_id_list  = ["sts.amazonaws.com"]
   thumbprint_list = [data.tls_certificate.oidc.certificates[0].sha1_fingerprint]
 }
 
 # Common assume role policies using OIDC federation
-locals {
-  oidc_provider_url_without_schema = replace(aws_iam_openid_connect_provider.oidc.url, "https://", "")
-}
-
 data "aws_iam_policy_document" "oidc_assume_role_policy" {
   statement {
     actions = ["sts:AssumeRoleWithWebIdentity"]
@@ -204,51 +207,22 @@ resource "aws_iam_role_policy" "hcp_cli" {
         Sid    = "EC2"
         Effect = "Allow"
         Action = [
-          "ec2:CreateDhcpOptions",
-          "ec2:DeleteSubnet",
-          "ec2:ReplaceRouteTableAssociation",
-          "ec2:DescribeAddresses",
-          "ec2:DescribeInstances",
-          "ec2:DeleteVpcEndpoints",
-          "ec2:CreateNatGateway",
-          "ec2:CreateVpc",
-          "ec2:DescribeDhcpOptions",
-          "ec2:AttachInternetGateway",
-          "ec2:DeleteVpcEndpointServiceConfigurations",
-          "ec2:DeleteRouteTable",
-          "ec2:AssociateRouteTable",
-          "ec2:DescribeInternetGateways",
-          "ec2:DescribeAvailabilityZones",
-          "ec2:CreateRoute",
-          "ec2:CreateInternetGateway",
-          "ec2:RevokeSecurityGroupEgress",
-          "ec2:ModifyVpcAttribute",
-          "ec2:DeleteInternetGateway",
-          "ec2:DescribeVpcEndpointConnections",
-          "ec2:RejectVpcEndpointConnections",
-          "ec2:DescribeRouteTables",
-          "ec2:ReleaseAddress",
-          "ec2:AssociateDhcpOptions",
-          "ec2:TerminateInstances",
-          "ec2:CreateTags",
-          "ec2:DeleteRoute",
-          "ec2:CreateRouteTable",
-          "ec2:DetachInternetGateway",
-          "ec2:DescribeVpcEndpointServiceConfigurations",
-          "ec2:DescribeNatGateways",
-          "ec2:DisassociateRouteTable",
-          "ec2:AllocateAddress",
-          "ec2:DescribeSecurityGroups",
-          "ec2:RevokeSecurityGroupIngress",
-          "ec2:CreateVpcEndpoint",
-          "ec2:DescribeVpcs",
-          "ec2:DeleteSecurityGroup",
-          "ec2:DeleteDhcpOptions",
-          "ec2:DeleteNatGateway",
-          "ec2:DescribeVpcEndpoints",
-          "ec2:DeleteVpc",
-          "ec2:CreateSubnet",
-          "ec2:DescribeSubnets"
+          "ec2:CreateDhcpOptions", "ec2:DeleteSubnet", "ec2:ReplaceRouteTableAssociation",
+          "ec2:DescribeAddresses", "ec2:DescribeInstances", "ec2:DeleteVpcEndpoints",
+          "ec2:CreateNatGateway", "ec2:CreateVpc", "ec2:DescribeDhcpOptions",
+          "ec2:AttachInternetGateway", "ec2:DeleteVpcEndpointServiceConfigurations",
+          "ec2:DeleteRouteTable", "ec2:AssociateRouteTable", "ec2:DescribeInternetGateways",
+          "ec2:DescribeAvailabilityZones", "ec2:CreateRoute", "ec2:CreateInternetGateway",
+          "ec2:RevokeSecurityGroupEgress", "ec2:ModifyVpcAttribute", "ec2:DeleteInternetGateway",
+          "ec2:DescribeVpcEndpointConnections", "ec2:RejectVpcEndpointConnections",
+          "ec2:DescribeRouteTables", "ec2:ReleaseAddress", "ec2:AssociateDhcpOptions",
+          "ec2:TerminateInstances", "ec2:CreateTags", "ec2:DeleteRoute", "ec2:CreateRouteTable",
+          "ec2:DetachInternetGateway", "ec2:DescribeVpcEndpointServiceConfigurations",
+          "ec2:DescribeNatGateways", "ec2:DisassociateRouteTable", "ec2:AllocateAddress",
+          "ec2:DescribeSecurityGroups", "ec2:RevokeSecurityGroupIngress", "ec2:CreateVpcEndpoint",
+          "ec2:DescribeVpcs", "ec2:DeleteSecurityGroup", "ec2:DeleteDhcpOptions",
+          "ec2:DeleteNatGateway", "ec2:DescribeVpcEndpoints", "ec2:DeleteVpc",
+          "ec2:CreateSubnet", "ec2:DescribeSubnets",
         ]
         Resource = "*"
       },
@@ -256,48 +230,31 @@ resource "aws_iam_role_policy" "hcp_cli" {
         Sid    = "ELB"
         Effect = "Allow"
         Action = [
-          "elasticloadbalancing:DeleteLoadBalancer",
-          "elasticloadbalancing:DescribeLoadBalancers",
-          "elasticloadbalancing:DescribeTargetGroups",
-          "elasticloadbalancing:DeleteTargetGroup"
+          "elasticloadbalancing:DeleteLoadBalancer", "elasticloadbalancing:DescribeLoadBalancers",
+          "elasticloadbalancing:DescribeTargetGroups", "elasticloadbalancing:DeleteTargetGroup",
         ]
         Resource = "*"
       },
       {
-        Sid    = "IAMPassRole"
-        Effect = "Allow"
-        Action = "iam:PassRole"
-        Resource = [
-          "arn:aws:iam::*:role/*-worker",
-          "arn:aws:iam::*:role/*-node-pool"
-        ]
+        Sid      = "IAMPassRole"
+        Effect   = "Allow"
+        Action   = "iam:PassRole"
+        Resource = ["arn:aws:iam::*:role/*-worker", "arn:aws:iam::*:role/*-node-pool"]
         Condition = {
-          "ForAnyValue:StringEqualsIfExists" = {
-            "iam:PassedToService" = "ec2.amazonaws.com"
-          }
+          "ForAnyValue:StringEqualsIfExists" = { "iam:PassedToService" = "ec2.amazonaws.com" }
         }
       },
       {
         Sid    = "IAM"
         Effect = "Allow"
         Action = [
-          "iam:CreateInstanceProfile",
-          "iam:DeleteInstanceProfile",
-          "iam:GetRole",
-          "iam:UpdateAssumeRolePolicy",
-          "iam:GetInstanceProfile",
-          "iam:TagRole",
-          "iam:RemoveRoleFromInstanceProfile",
-          "iam:CreateRole",
-          "iam:DeleteRole",
-          "iam:PutRolePolicy",
-          "iam:AddRoleToInstanceProfile",
-          "iam:CreateOpenIDConnectProvider",
-          "iam:ListOpenIDConnectProviders",
-          "iam:DeleteRolePolicy",
-          "iam:UpdateRole",
-          "iam:DeleteOpenIDConnectProvider",
-          "iam:GetRolePolicy"
+          "iam:CreateInstanceProfile", "iam:DeleteInstanceProfile", "iam:GetRole",
+          "iam:UpdateAssumeRolePolicy", "iam:GetInstanceProfile", "iam:TagRole",
+          "iam:RemoveRoleFromInstanceProfile", "iam:CreateRole", "iam:DeleteRole",
+          "iam:PutRolePolicy", "iam:AddRoleToInstanceProfile",
+          "iam:CreateOpenIDConnectProvider", "iam:ListOpenIDConnectProviders",
+          "iam:DeleteRolePolicy", "iam:UpdateRole", "iam:DeleteOpenIDConnectProvider",
+          "iam:GetRolePolicy",
         ]
         Resource = "*"
       },
@@ -305,14 +262,9 @@ resource "aws_iam_role_policy" "hcp_cli" {
         Sid    = "Route53"
         Effect = "Allow"
         Action = [
-          "route53:ListHostedZonesByVPC",
-          "route53:CreateHostedZone",
-          "route53:ListHostedZones",
-          "route53:ChangeResourceRecordSets",
-          "route53:ListResourceRecordSets",
-          "route53:DeleteHostedZone",
-          "route53:AssociateVPCWithHostedZone",
-          "route53:ListHostedZonesByName"
+          "route53:ListHostedZonesByVPC", "route53:CreateHostedZone", "route53:ListHostedZones",
+          "route53:ChangeResourceRecordSets", "route53:ListResourceRecordSets",
+          "route53:DeleteHostedZone", "route53:AssociateVPCWithHostedZone", "route53:ListHostedZonesByName",
         ]
         Resource = "*"
       },
@@ -320,13 +272,10 @@ resource "aws_iam_role_policy" "hcp_cli" {
         Sid    = "S3"
         Effect = "Allow"
         Action = [
-          "s3:ListAllMyBuckets",
-          "s3:ListBucket",
-          "s3:DeleteObject",
-          "s3:DeleteBucket"
+          "s3:ListAllMyBuckets", "s3:ListBucket", "s3:DeleteObject", "s3:DeleteBucket",
         ]
         Resource = "*"
-      }
+      },
     ]
   })
 }